[dataset] reorganize dataset modules
