+++ deeposlandia/datasets/__init__.py
@@ -1,0 +1,6 @@
+"""Dataset modules
+
+Each considered dataset is represented by its own module, and its own class
+that inherits from Dataset class (see `dataset.py`).
+
+"""
